requirements.txt
streamlit
pandas 
plotly.express 
plotly.figure_factory 
seaborn 
matplotlib.pyplot as plt
sklearn.model_selection import train_test_split
numpy 
sklearn.model_selection import train_test_split 
sklearn.impute import SimpleImputer, KNNImputer 
sklearn.preprocessing import OrdinalEncoder, LabelEncoder, OneHotEncoder 
category_encoders import BinaryEncoder # Nominal Encoding 
imblearn.under_sampling import RandomUnderSampler 
imblearn.over_sampling import RandomOverSampler, SMOTE 
sklearn.preprocessing import StandardScaler, MinMaxScaler, RobustScaler
